refactor: change type defintions for input elements in form
I explicitely typed the name properties of TextInput and TextArea to be a string value equivalent to the name I used in the corresponding interface in Form. This caused unneccesary coupling, so I changed the type definition since the two input components dont need to be concerend with the interface logic in Form.